MAJOR: added embedded-box coulomb truncation for fluids!
Specifically, fluids can be run in a box that is twice as large along 'truncated' directions as the electronic system. The fluid solvers are almost unmodified and not complicated in any way due to this.

Using this modification, solvation calculations (including full JDFT) can be run in very tight boxes, just enough to converge the KE of the explicit systems. Speed gains will obviously be most dramatic when the electronic system cost dominates.