Changed how preconditioning is handled in L-BFGS
Now using the preconditioner as a metric (consistent with CG),
rather than as an initial Hessian (in the original BFGS papers).
This allows caching of the preconditioned gradient, since the
preconditioner is always applied to the gradient at a point on the
minimization trajectory. This is a must for using L-BFGS with the
electronic system, where precondition is not simply a matrix multiply
due to the nontrivial handling of  fillings / unoccupied state gradients.