Converted linminCubic to a weak Wolfe-criterion terminated step.
Basically, if the test step reduces the energy and gdotd enough,
it is accepted as the real step. This can lead to 1.5-2x speedup
relative to the more conservative linminQuad, but the inaccurate
linmin can only be tolerated by L-BFGS (and not by CG).

Updated the linminMethod mechanism selection, whereby the
dirUpdateScheme can control which linmin is default eg.
Quad for CG and CubicWolfe for BFGS.