Fixed bug in lattice minimization with SCF
(The bug was introduced in the recent commit that eliminated unnecessary orthogonalizations in elecEnergyAndGrad in SCF mode.
This caused an error in the finite difference lattice gradient calculation as the wavefunctions are not correctly reorthonormalized in the test steps.)

Also fixed lattice minimization with SCF for metals which was probably broken even before aforementioned commit, due to incorrect handling of fillings / aux Hamiltonian.