Proposing a fix for the grad_K -> NAN premature termination in FermiFillingsAux minimization: each time grad_K is about to become NAN, run a band structure solve at the current potential to converge the empty states (the indefiniteness in the preconditioner is primarily because of mixing between filled and empty state gradients (I think)).
Seems to work well, but for the following caveats:

* It could increase run time: by O(10) elec steps for each time the issue is detected.
Make sure elec-eigen-algo is set to Davidson (the default), the time penalty is likely
worse for CG bandstructure.

* It could increase memory consumption. Not as much of an issue when number of kpoints
per process is large, since the band structure is done one kpoint at a time. If this
becomes too much of an issue, we could solve it in an extremely messy manner by
somehow getting the minimizer to deallocate its gradients (prefer not to).