Fixed mysterious phases in transformation of relativistic spin-angle functions under rotations.
These phases had to be introduced during the relativistic DFT+U implementation. TUrns out it was a complex conjugacy error, and the standard Pauli matrix transformations are adequate after fixing this.

Added spinor transformations to Wannier so as to support symmetries in relativistic mode (obviously this needed the above fix).