Added nonlocal corrections to the CEDA plasma frequency sum rule.
This corrects the free electron plasma frequency with a commutator of
the cross-k overlap operator and the Hamiltonian (contributions from
NLpsp and DFT+U) and generalizes the result of PRB 78, 085125 (2008)
to be exact for general pseudopotentials!

Almost no change for aluminium and seems to work very well for noble metals
(pending a more expensive test on full momentum grid).