Started the migration of command documentation to Doxygen
Added complete list of commands to the top level documentation page, but most of the subpages are missing and the links therefore point to the corresponding command class documentation instead.

For each command, need to move the comments and format strings into a doxygen block. Finished processing only the commands in coulomb_interaction.cpp. As we continue the porting, the doxy documentation will gradually fill in while jdftx -t and the older html manual gradually empties out.