Enabled Kerker mixing for potential-mixed SCF.
This makes the mixing more stable in general, but i doing so, slows down the convergence of "easy" cases.

Commiting this modification because the old behavior can be recovered by setting qKerker = 0 in the input file (while the reverse is not true)

Need to update default settings of SCF after systematic convergence testing.