Fixed coordinate system error for symmetrization in IonicMinimizer::constrain(). Also moved ion constraints from the preconditioning logic in compute() to constrain(), which is of course more logical.
Thsi function mistakenly passed Cartesian forces to the symmetrization routine that expected contravariant lattice coordinates

This bug has existed since the dawn of time (i.e. when JDFTx first appeared on SourceForge),
but never mattered because IonicMinimizer::constrain was never called except during fdTest(),
and even then it may have been noticeable only for calculations with non-orthogonal axes
(most of the ion-min fdtests were done on molecules in cubic boxes).

It became critical in 1.2.0 when constrain was called on the minimization search direction
due to an unrelated reason: that was needed for the Y-less electronic minimization logic.